🚚 Move floating point arbitraries into /arbitrary
