+++ src/fast-check-default.ts
@@ -44,7 +44,8 @@ import { Arbitrary } from './check/arbitrary/definition/Arbitrary';
 import { Shrinkable } from './check/arbitrary/definition/Shrinkable';
 import { dictionary } from './arbitrary/dictionary';
 import { emailAddress } from './arbitrary/emailAddress';
-import { double, float, DoubleConstraints, FloatConstraints } from './check/arbitrary/FloatingPointArbitrary';
+import { double, DoubleConstraints } from './arbitrary/double';
+import { float, FloatConstraints } from './arbitrary/float';
 import { frequency, WeightedArbitrary, FrequencyValue, FrequencyContraints } from './arbitrary/frequency';
 import { compareBooleanFunc } from './arbitrary/compareBooleanFunc';
 import { compareFunc } from './arbitrary/compareFunc';
@@ -144,7 +145,7 @@ import { ArbitraryWithContextualShrink } from './check/arbitrary/definition/Arbi
 import { CommandsContraints } from './check/model/commands/CommandsContraints';
 import { PreconditionFailure } from './check/precondition/PreconditionFailure';
 import { RandomType } from './check/runner/configuration/RandomType';
-import { FloatNextConstraints } from './check/arbitrary/FloatNextArbitrary';
+import { FloatNextConstraints } from './arbitrary/_next/floatNext';
 import { int8Array, IntArrayConstraints } from './arbitrary/int8Array';
 import { int16Array } from './arbitrary/int16Array';
 import { int32Array } from './arbitrary/int32Array';
@@ -155,7 +156,7 @@ import { uint32Array } from './arbitrary/uint32Array';
 import { float32Array, Float32ArrayConstraints } from './arbitrary/float32Array';
 import { float64Array, Float64ArrayConstraints } from './arbitrary/float64Array';
 import { sparseArray, SparseArrayConstraints } from './arbitrary/sparseArray';
-import { DoubleNextConstraints } from './check/arbitrary/DoubleNextArbitrary';
+import { DoubleNextConstraints } from './arbitrary/_next/doubleNext';
 import { NextArbitrary } from './check/arbitrary/definition/NextArbitrary';
 import { NextValue } from './check/arbitrary/definition/NextValue';
 import { convertFromNext, convertFromNextWithShrunkOnce, convertToNext } from './check/arbitrary/definition/Converters';
