+++ test/unit/check/arbitrary/FloatingPointArbitrary.spec.ts
@@ -2,7 +2,8 @@ import * as fc from '../../../../lib/fast-check';
 
 import { Arbitrary } from '../../../../src/check/arbitrary/definition/Arbitrary';
 import { Random } from '../../../../src/random/generator/Random';
-import { float, double } from '../../../../src/check/arbitrary/FloatingPointArbitrary';
+import { double } from '../../../../src/arbitrary/double';
+import { float } from '../../../../src/arbitrary/float';
 
 import * as genericHelper from './generic/GenericArbitraryHelper';
